src/requirements.txt
asgiref==3.2.10
attrs==20.2.0
autobahn==20.7.1
Automat==20.2.0
certifi==2020.6.20
cffi==1.14.3
channels==2.4.0
chardet==3.0.4
cli-helpers==2.1.0
click==7.1.2
configobj==5.0.6
constantly==15.1.0
coverage==5.3
cryptography==3.1.1
daphne==2.5.0
defusedxml==0.6.0
diff-match-patch==20200713
Django==3.0.10
django-3-jet==1.0.8
django-ckeditor==6.0.0
django-classy-tags==2.0.0
django-cms==3.7.4
django-filter==2.4.0
django-formtools==2.2
django-import-export==2.4.0
django-js-asset==1.2.2
django-sekizai==2.0.0
django-tagging==0.5.0
django-taggit==1.3.0
django-treebeard==4.3.1
djangocms-admin-style==2.0.0
djangorestframework==3.12.1
djangorestframework-jwt==1.11.0
easy-thumbnails==2.7
et-xmlfile==1.0.1
flake8==3.8.3
gunicorn==20.0.4
humanize==3.0.1
hyperlink==20.0.1
idna==2.10
importlib-metadata==2.0.0
incremental==17.5.0
iniconfig==1.0.1
isort==4.3.21
jdcal==1.4.1
Markdown==3.3
MarkupPy==1.14
mccabe==0.6.1
odfpy==1.4.1
openpyxl==3.0.5
packaging==20.4
pgcli==3.0.0
pgspecial==1.11.10
Pillow==7.2.0
pluggy==0.13.1
prompt-toolkit==3.0.7
psycopg2==2.8.6
ptvsd==4.3.2
py==1.9.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.6.0
pycparser==2.20
pyflakes==2.2.0
Pygments==2.7.1
PyHamcrest==2.0.2
PyJWT==1.7.1
pyOpenSSL==19.1.0
pyparsing==2.4.7
pytest==6.1.1
pytest-testmon==1.0.3
python-decouple==3.3
pytz==2020.1
PyYAML==5.3.1
requests==2.24.0
service-identity==18.1.0
setproctitle==1.1.10
six==1.15.0
sqlparse==0.3.1
tablib==2.0.0
tabulate==0.8.7
terminaltables==3.1.0
toml==0.10.1
Twisted==20.3.0
txaio==20.4.1
urllib3==1.25.10
wcwidth==0.2.5
xlrd==1.2.0
xlwt==1.3.0
zipp==3.3.0
zope.interface==5.1.2
pytest-django==4.1.0
pytest-cov==2.10.1
pytest-factoryboy==2.0.3
boto3==1.16.18
django-storages==1.10.1
django-q==1.3.4
django-redis==4.12.1
django-widget-tweaks
xlrd2==1.2.6
django-ckeditor==6.0.0